feat: improve arhitecture of the application with modules
