+++ assets/scripts/aggregate.js
@@ -1,0 +1,11 @@
+
+export { create } from './create.js';
+export { header } from './header.js';
+export { createCard } from './createCard.js';
+export { cart, deleteItem, restoreFromLocalStorage,  mirrorToLocalStorage, displayItems, displayTotal, addToCart } from './cart.js';
+export { addStyles } from './addStyles.js';
+export { addModalShowMore, showMoreModal } from './modal.js';
+export { dragAndDrop } from './dragAndDrop.js';
+export { footer } from './footer.js';
+export { main } from '../../pages/index.js';
+export { checkDate, validateCheckbox, changeSubmitBtn, checkValidityInputs, summarizeOrder } from './formValidation.js';
