+++ pages/index.html
@@ -1,16 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
-   </head>
-<body>
+  </head>
+  <body>
     <main class="wrapper" id="wrapper"></main>
-
-    <script type="text/javascript" src="data.js"></script>
-    <script type="text/javascript" src="main.js"></script>
-    
-</body>
+    <script type="module" src="index.js"></script>
+  </body>
 </html>
